Clean up active site parsing logic (#110)
