+++ src/main/java/org/moltimate/moltimatebackend/parser/ActiveSiteParser.java
@@ -1,0 +1,10 @@
+package org.moltimate.moltimatebackend.parser;
+
+import org.moltimate.moltimatebackend.model.ActiveSite;
+
+import java.util.List;
+
+public interface ActiveSiteParser {
+
+    List<ActiveSite> parseMotifs();
+}
